JAVA-2507: Move Base64 support from driver-core module to bson module, and make it public
